particle_system part 4
issue with .obj remains.
Separated the highlighter into its own module, and incorporated it into the parameters.txt portion of convert. Now what's missing is connect it in the Blender API. Almost there.
