+++ .vs/VSWorkspaceState.json
@@ -1,8 +1,8 @@
 {
   "ExpandedNodes": [
     "",
-    "\\gui"
+    "\\scripts"
   ],
-  "SelectedNode": "\\gui",
+  "SelectedNode": "\\scripts\\Main_Body.py",
   "PreviewInSolutionExplorer": false
 }
