+++ scripts/parameters.txt
@@ -1,11 +1,13 @@
 .com
-C:/Documents/Gaussian-2-Blender/input_examples/com_files/organic
-acrylonitrile.com
+C:/Documents/Gaussian-2-Blender/input_examples/com_files/animations/water
+water_frame1.com
 C:\Documents\Gaussian-2-Blender\output\
-acrylonitrile
+water_frame1
 Ball-and-Stick
-.obj
+.fbx
 0
 0
 ---
 0
+O01, H02
+O01-H02
